test: refactor driver
