+++ jina/excepts.py
@@ -181,3 +181,7 @@ class ImageAlreadyExists(Exception):
 
 class BadFlowYAMLVersion(Exception):
     """ Exception when Flow YAML config specifies a wrong version number"""
+
+
+class LookupyError(Exception):
+    """Base exception class for all exceptions raised by lookupy"""
